requirements.txt
python-telegram-bot
confi
requests
pysocks